Fix imports of refactored typedefs
